requirements.txt
uvicorn
fastapi[all]
pyppeteer
colorgram.py
flake8
